MootCourt/agent_stuff/requirement.txt
# CSV to Vector Database - Essential Requirements Only

# Core LangChain
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# Google Gemini LLM Integration
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0

# Vector Database
chromadb>=0.4.22

# Embeddings (Free, Local)
sentence-transformers>=2.2.2
torch>=2.0.0

# Data Processing
pandas>=2.0.0

# Environment Variables
python-dotenv>=1.0.0